pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0", "hatch-vcs>=0.3.0"]
build-backend = "hatchling.build"

[project]
name = "jurisai"
description = "Legal Assistant for Ghana and Sierra Leone"
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE" }
authors = [
    { name = "Saikrishna Paila", email = "spaila@worldbank.org" }
]
dynamic = ["version"]
python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Government",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Legal",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "streamlit>=1.24.0",
    "fastapi>=0.68.0",
    "langchain>=0.0.200",
    "openai>=0.27.0",
    "pinecone-client>=2.2.1",
    "google-cloud-translate>=3.8.1",
    "python-dotenv>=0.19.0"
]

[project.optional-dependencies]
docs = [
    "docutils==0.17.1",
    "jupyter-book>=1,<2",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
    "pre-commit>=3.0",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/**/*"
]

[tool.hatch.version]
source = "vcs"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3 